tinc-tailor: default properties from DEFAULT section
this exposes the default values in Tailor.properties, which can be used
from argparse and run.  Setting an unset variable will set it for the
individual hosts. However *changing* a variable at this point will not
cause it to be changed for the individual hosts because ConfigParser will
use the original value to override!